% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_index.R
\name{boot_index}
\alias{boot_index}
\title{Bootsrapped randomly samples values}
\usage{
boot_index(data, id, value, reps = 1000)
}
\arguments{
\item{data}{a data.frame}

\item{id}{variable name, to be aggregated around}

\item{value}{variable name, contains the value to take mean over}

\item{reps}{Number of reps, defaults to 1000}
}
\value{
a data.frame
}
\description{
Bootsrapped randomly samples values
}
