% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ii_summarize.R
\name{ii_summarize}
\alias{ii_summarize}
\title{Aggregated}
\usage{
ii_summarize(data, id, value, fun = mean, ...)
}
\arguments{
\item{data}{a data.frame}

\item{id}{variable name, to be aggregated around}

\item{value}{variable name, contains the value to take mean over}

\item{fun}{function, summary statistic function  to be calculated. Defaults
to `mean`.}

\item{...}{Additional arguments for the function given to the argument fun.}
}
\value{
data.frame
}
\description{
Aggregated
}
