% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_index.R
\name{bootter}
\alias{bootter}
\title{Bootsrapped randomly samples values}
\usage{
bootter(data, data_full, value, id, reps)
}
\arguments{
\item{data}{a data.frame with data summaries by a variable.}

\item{data_full}{a data.frame which is the unsummaried version of data.}

\item{value}{variable name, contains the value to take mean over}

\item{id}{variable name, to be aggregated around}

\item{reps}{Number of reps, defaults to 1000}
}
\value{
a data.frame
}
\description{
Bootsrapped randomly samples values
}
