% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ii_summarize.R
\name{ii_summarize}
\alias{ii_summarize}
\title{Aggregated}
\usage{
ii_summarize(data, id, value)
}
\arguments{
\item{data}{a data.frame}

\item{id}{variable name, to be aggregated around}

\item{value}{variable name, contains the value to take mean over}
}
\value{
data.frame
}
\description{
Aggregated
}
