% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{meth_bootstrap}
\alias{meth_bootstrap}
\title{Bootstrapped randomly samples values}
\usage{
meth_bootstrap(data, reps, method = c("perm_v1", "perm_v2", "perm_v3"))
}
\arguments{
\item{data}{a methcon data.frame output from ii_sumarize().}

\item{reps}{Number of reps, defaults to 1000}

\item{method}{Character, determining which method to use. See details for
information about methods. Defauls to "perm_v1".}
}
\value{
a data.frame
}
\description{
"perm_v1" (the default method) will sample the variables the rows
independently. "perm_v2" will sample regions of same size while allowing
overlap between different regions. "perm_v3" will sample regions under the
constraint that all sampled regions are contained in the region they are
sampled in.
}
